Modify `Features` class
